requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.2
colorama==0.4.6
decorator==5.2.1
executing==2.2.1
fastapi==0.136.0
h11==0.16.0
idna==3.10
importlib_metadata==9.0.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
mariadb==1.1.12
matplotlib-inline==0.2.1
mysql-connector-python==9.6.0
packaging==25.0
parso==0.8.6
prompt_toolkit==3.0.52
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pyspark==3.5.5
requests==2.32.3
setuptools==82.0.1
stack-data==0.6.3
starlette==1.0.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
uvicorn==0.44.0
watermark==2.6.0
wcwidth==0.6.0
zipp==3.23.0
